QDoc: Move the catch dependency under src
QDoc uses the Catch2 testing-framework for some of its tests.

The dependency is included as a header-only library that lived under
"tests/auto/qdoc/".

Due to a certain restructuring that is happening in QDoc, with one of the
goals, among others, being to increase the locality of QDoc-related code
under "src/qdoc", the Catch2 dependency is now moved under "src/qdoc".

To allow tests that depend on it to keep their usages without requiring
relative paths from the "tests" directory to the "src" directory, a
library target, "Qt::QDocCatchPrivate" was created for the dependency.

The target can be linked to to have access to the Catch2 dependency and
the Catch2 header can be included with "catch/catch.hpp".

To allow for this specific include path to work, the internal directory
structure of the Catch2 dependency was slightly modified, adding some
additional intermediate directories in between its root and the library
header.

Tests under "tests/auto/qdoc" that used the dependency now link to it
instead. Hence, their "CMakeLists.txt" files were modified to avoid
using the dependency as an include directory and instead use
"Qt::QDocCatchPrivate" as a library.

The relevant inclusion of the Catch2 header in those sources was
modified to respect the new "catch/catch.hpp" path.

Change-Id: Ifd462f582d197f78a8ffbbbe7b2cf9ae25dbc991
Reviewed-by: Paul Wicking <paul.wicking@qt.io>