Doc: Let QDoc manual reflect new example guidelines
The new guidelines for examples in Qt recommend one avoids use of
the word "Example" in an example's name. The QDoc manual contains
documentation for the example manifest files generated by QDoc.
This documentation references content generated with a naming
scheme that doesn't adhere to the new guidelines. This patch
removes the use of the word 'example' from example names such that
example content in the QDoc manual reflects the new example naming
scheme in Qt.

Fixes: QTBUG-113393
Change-Id: Idd95bd06c6b67471aed5cb6147386235ceac79c7
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>