QDoc: Ensure CMD_BRIEF works with line comments
QDoc's `CMD_BRIEF` takes a paragraph as its argument. This paragraph may
contain any number of QDoc commands or macros. These commands or macros
will, in some cases, consume the first newline character that follows.
As `CMD_BRIEF` relies on two newline characters as its end condition,
this may cause the command to erroneously include content that follows
as part of its argument.

This misbehavior has previously been rectified for a range of QDoc's
commands. However, the case for line comments has yet to be addressed.
This patch extends the generated output test by adding test data that
combines `CMD_BRIEF` with QDoc line comments. Of special significance
is the following;

- The new page `page-with-comment-after-brief.html` exposes how QDoc
  line comments are consumed by `CMD_BRIEF` when they immediately
  follow the command.
- The new page `page-with-comment-in-brief.html` exposes how QDoc line
  comments break `CMD_BRIEF` if the case for handling line comments in
  `DocParser::parse` sees the addition of a call to `leavePara()`,
  which is the workaround that has been put in place in other locations
  that expose the issue.
- The warning about a missing full stop for briefs when generating
  html output is not generated if a `CMD_BRIEF` encounters a line
  comment. However, multiple line comments are not a problem.

To rectify the misbehavior in the case of line comments, let the parsing
of line comments move back to the newline character, '\n', before
breaking the switch that's responsible for dispatch. This ensures that
`CMD_BRIEF` will process its argument correctly and not end prematurely.

As a side effect of this, QDoc no longer incorrectly reports about an
unknown 'sa' macro if there are line comments as part of a see also
command, as reported in a separate issue.

Task-number: QTBUG-105754
Fixes: QTBUG-113118
Change-Id: Ifee2a78193d99ae6315103c8862746108a69a74d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Luca Di Sera <luca.disera@qt.io>