QDoc: Allow line comments in see also command
If QDoc encounters a line comment (`//!`) when parsing a see also
command (`\sa`), QDoc attempts to link to the comment marker and each
word in the comment. This generates a range of warnings and may be
confusing for the user.

This patch adds reproducing test data to the generated output test. The
bug is resolved by letting `DocParser::parseAlso` recognize line
comments and skipping the remainder of the line in case a line comment
is encountered. Finally, missing documentation for the method is added.

Fixes: QTBUG-105754
Change-Id: I2f8aeaf6206c7cf8da8a2a16613ef53213dec460
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>