QDoc: Remove Generator::generateAtom/2
`Generator`, the base class that takes care of the process of producing
the final output documentation, provides a series of virtual methods
that its child classes can override to take care of formatting the
output with their format-specific concerns.

Many of those methods provides two overloads, with the only difference
between the two being that whether a `CodeMarker` is passed to the
method.

A `CodeMarker` is used to "format" certain code elements, such as C++
code, on the generated documentation.

Of the two main formats that QDoc uses, "codemarking" is required by
HTML and unused by DocBook, as DocBook provides its own syntax
highlighting.

Hence, the non `CodeMarker` overload is used by the `DocBookGenerator`
class, while the `CodeMarker` overload is used by the `HtmlGenerator`
class.

For some of those overloaded methods, the implementation is mostly about
structural induction, with the actual implementation shareable between
formats.

Such an example is, `Generator::generateAtomList`, which takes care of
generating the output documentaton for a list of `Atom`s, the
intermediate representation that QDoc uses for user-provided
documentation.

When this is the case, if the implementation uses an overloaded method
that is overridden, it slightly diverges due the fact that the called
method differs between the child classes implementation, thus making it
difficult to condense the code into a single shared method.

To avoid this issue for `Generator::generateAtomList` specifically, the
non `CodeMarker` overload of `Generator::generateAtom`, which is
delegated to by `generateAtomList`, is now removed.

The overridden `DocBookGenerator::generateAtom` was modified to align
to the `CodeMarker` overload.
Calls to `DocBookGenerator::generateAtomList` were modified to received
a `nullptr` `CodeMarker`, while the implementation was modified to
ignore the `CodeMarker` argument.

While similar overloads exist, which might or might not produce the same
diversion, they are for now left untouched and are intended to be
considered, in the short term, only on a "per-case" basis if they are
an impediment to unify-able methods.

Change-Id: Icf98da5bf6d64fb90098d1a25d6711f969e68e25
Reviewed-by: Paul Wicking <paul.wicking@qt.io>