QDoc: Generate correct macro for DocBook format
QDoc allows the definition of code-expanding macros in its
configuration.

A macro can be defined with a format-specific expansion, so that the
same macro can be used for different formats.

For example, a macro `para` might expand to "<p>" when QDoc is producing
an HTML-formatted output documentation and to "<para>" when QDoc is
producing a DocBook-formatted output documentation.

When QDoc expands the usage of a macro, it produces an expansion for
each format specific version of a macro.

Each expansion is protected by a `FormatIf` `Atom`, the internal
representation that QDoc uses for user-provided documentation, that
carries the name of the format that the macro is expanded for, so that
the later-performed generation can decide whether to generate the
expanded output or not based on the currently generated format.

Similarly, QDoc provides a command, "\raw", to inject user-provided code
in the output documentation.
"\raw" requires the user to provide a "format" argument, such as "HTML",
that represents the format of the code they are trying to inject into
the output.

A "\raw" command is represented, internally, the same as a macro
expansion, using a `FormatIf` `Atom` that carries the name of the format
that the user provided, so that the latter generation phase can filter
out the instances of the commands that are unrequired when generating a
certain format.

During the generation phase, this "filtering" process is performed by
`Generator::generateAtomList`, which traverses the `Atom` representation
taking into account whether a format-specific macro expansion should be
generated or not.

Child classes of `Generator`, the classes that implement the logic for
format-specific output generation, generally delegate the traversal of
Atoms to `Generator::generateAtomList`, thus correctly handling the
presence of `FormatIf` atoms.

The `DocBookGenerator` child class of `Generator`, which takes care of
generating the output document in the DocBook format, reimplements,
instead, a slightly different version of `generateAtomList`, due to
trivial interface differences.

The implemented version diverged from the initial `Generator`
implementation, lacking the correct handling of `FormatIf` atoms.

Hence, when presented with a series of macro expansions in the form of a
series of `FormatIf` chains, `DocBookGenerator::generateAtomList` would
always produce the expansion of the first encountered `FormatIf` in the
chain, independently of format.

This has not generally been a problem. If the first expansion is always
the DocBook-specific expansion, `DocBookGenerator::GenerateAtomList`
will expand macros "correctly".

Due to an implementation detail, when format-specific macros are
expanded, their expansions are ordered lexicographically based on the
format name.

As the main formats that QDoc supports are DocBook and HTML, most
format-specific macros that Qt's documentation uses are in those two
formats.

Based on the above mentioned implementation detail, it so happened that
the `DocBook` expansion has always been the first presented
format-specific expansion, so that the issue never manifested itself.

Nonetheless, any project that presents a format-specific macro whose
format name is a lexicographic predecessor of `DocBook`, such as "ABC",
will produce incorrect output when the `DocBook`-formatted documentation
is generated.

To avoid the issue, `DocBookGenerator::generateAtomList` is now modified
to ensure that a `FormatIf` `Atom` is produced only when its associated
format is `DocBook`.

`generateAtomList` uses a boolean argument, `generate`, to decide
whether a chain of `Atom`s should be produced or not.
This is used when `generateAtomList` recurses to skip `FormatIf` atoms
whose expansions should not be generated.

Before, `generateAtomList` would always pass `false` when recurring on a
`FormatIf` that was not the first in its chain.
This is now changed so that `DocBookGenerator::generateAtomList` will
condition the value of `generate` in recursive calls based on whether
the currently handled `FormatIf` `Atom` represent an expansion in the
DocBook format or not.

To do so, the already existing `Generator::canHandleFormat`, which
predicates whether the currently used `Generator`-based class matches
the format name passed to it, was used.

Thus, `DocBookGenerator::generateAtomList` will now avoid expanding
format-specific macros to their non-DocBook expansion.

Similarly, it will now avoid injecting "\raw" code for a format that is
not "DocBook".

The new implementation closely follows the one in
`Generator::generateAtomList`, to align the `DocBookGenerator` specific
version to the on in `Generator`.

Due to the presence of this bug, some amount of code was written to
convert the known cases of "\raw HTML" commands that were erroneously
processed by "DocBookGenerator".
The code is now removed as dead code and the behavior is modified so
that "DocBookGenerator" simply writes the user provided code to the
output documentation, as per the expected behavior.

`DocBookGenerator::m_rewritingCustomQmlModuleSummary`, a private member
that was only used in removed code, is now removed.

`DocBookGenerator::writeRawHtml`, which was only used in the removed
code, is removed as it is now dead code.

The removal of the code will prevent certain parts of the documentation
to be generated, while it previously was thanks to the conversion. Those
cases where a DocBook specific equivalent is possible will be handled
separately from this patch by modifying the relevant documentation.

QDoc's regression tests were regenerated to respect the fixed behavior
for the DocBook format.

The documentation for the "\raw" command was updated to mention that the
user-provided format must be the same as the one that QDoc is generating
for the command to have an effect.

Change-Id: If4ee2ec8bfb96ccb7b81d17fbbcfd917f8b51f15
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>