Doc: Remove raw listings from QDoc manual
QDoc provides the "\raw" command to allow user to inject code in the
generated output that sidesteps QDoc generation rules.

The "\raw" command is intended for special cases that QDoc currently
does not support. Furthermore, the "\raw" command allows for
format-specific customization.

In the QDoc manual, there's extended use of `\raw HTML` for the purpose
of displaying possible output from QDoc's commands. This content
conflicts with QDoc as a multi-format generator, which becomes apparent
when generating e.g. DocBook format. The content isn't well maintained,
and may give the incorrect impression that QDoc is a generator of HTML,
exclusively. Adding DocBook output for the range of output examples is
understood as introducing significant maintenance overhead. Mixing the
formats (e.g. replacing a portion of the current HTML output with its
DocBook equivalent) is considered detrimental to the reader experience,
as it may become confusing.

To avoid setting precedence by adding example DocBook content as an
addition to the HTML content that is currently there, remove rendered
output from QDoc's reference manual.

Change-Id: I22b456a7465d4f9dd8e219f5ecbb3523fd03ebec
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>