QDoc: Remove `DocBookGenerator::generateAtomList`
When QDoc parses the user provided documentation, it converts into its
own intermediate representation, a linear collection of `Atom`s, which
is later traversed and processed when QDoc is generating the final
output documentation.

The `Generator` class family takes care of performing this process so as
to generate the output of a QDoc run.
In particular, `Generator::generateAtomList` performs the structural
traversal of the linear collection of `Atom`s that represents the
user-defined documentation for a project.

While this traversal is common to all generators, `DocBookGenerator`,
which takes care of producing output in the "DocBook" format, implements
its own version of the method, which is semantically equivalent to the
original `Generator` implementation.

The method was implemented due to a small difference in API design.
Most methods in `Generator` will use a `CodeMarker`, an internal family
of classes used to lex code snippets provided in
the documentation to provide highlighting or embellishments.
The "DocBook" format provides this through other means, so that a
`CodeMarker` is not required when producing its output documentation.

To avoid having to pass a "nulled" `CodeMarker`, the API for `Generator`
was split into "non-CodeMarker" and "CodeMarker" methods, so that it
diverged, and required the custom implementation that `DocBookGenerator`
provides.

The custom implementation was originally meant to follow the one in
`Generator` but fell out of sync during the years, resulting in the
presence of certain bugs.

Recent patches fixed those bugs and aligned the implementation to the
one in `Generator`, making them interchangeable.

Thus, the custom implementation of `generateAtomList` for
`DocBookGenerator` was removed in favor of implicitly calling the one
from `Generator`.

The declaration and definition of `DocbBookGenerator::generateAtomList`
were removed from the respective header and implementation files.

Furthermore, the single call of `generateAtomList` in `DocBookGenerator`
was aligned to the new interface by passing a "nulled" `CodeMarker`.

Change-Id: I9aedc5a30a2cbe44da46ae30cd34a93a1910780a
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>