QDoc: Use platform dependent EOL characters
For a correctly configured QDoc project, QDoc generates one or multiple
files as output artifacts. These files are of various types depending on
the configured output format(s) (e.g. `.html`, `.xml`, `.webxml`, etc),
but are ultimately variations of files that contain plain UTF-8 encoded
text.

QDoc relies on `QFile` as the i/o device for output files for its
`Generator` base class and its subclasses. When opening a file for
writing, QDoc specifies `QFile::WriteOnly` as open mode for the target
file. `QFile`'s underlying mechanism to offer translation for
end-of-line (EOL) terminator characters is unused. The file object is
then wrapped in a `QTextStream`, and content in the form of `QString`
objects are piped to this stream to be written to the file.

This results in QDoc writing contents to the file verbatim, respecting
any whitespace character as just any other piece of data. In QDoc, the
EOL character sequence is `\n`, the new line character, which is common
for "*nix systems" (e.g. Linux, Darwin, BSD, etc). However, on Windows,
the EOL character sequence is, for historic reasons, `\r\n`. This may
cause issues when other tools, e.g. version control clients or text
editors, open the resulting file on Windows.

To mitigate this behavior in the `Generator` family of classes, this
patch modifies the method responsible for opening files for output,
`Generator::openSubPageFile`, by passing an additional flag,
`QFile::Text`, to the `QFile` constructor. This flag ensures that EOL
characters are translated to the local encoding. For Win32 platforms,
this means `\n` characters will be written as `\r\n`.

QDoc's `ManifestWriter`, `HelpProjectWriter`, and `IndexFiles` classes
already pass the aforementioned flag and are as such not affected by
this issue.

Fixes: QTBUG-55119
Change-Id: I4b381fd705f417cc974e470113023ffcb30475c2
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>