qdoc: Add capability to issue a warning for missing .qhp configuration
A frequent source of documentation build failures is a situation where
a documentation project has no help project configuration, but the
'qch_docs' build target expects the .qhp file to exist.

Allow the 'qhp' configuration variable to accept a boolean value;
if set to true, QDoc will issue a warning for missing .qhp config.

This can then be set in the shared global configuration in qtbase.
Together with a 'warninglimit' variable, a missing .qhp config
blocks integrations that would otherwise break the top-level
documentation build.

Fixes: QTBUG-114181
Change-Id: Ibb99bdc93d98dc980434f698bc7f8288acc88984
Reviewed-by: Paul Wicking <paul.wicking@qt.io>