QDoc: Generate deprecation warning in QML Type page headers
QDoc features a command, `\deprecated`, which is used to record whether
certain entities (e.g. C++ classes or QML types) are or contain APIs
that are considered deprecated. The status is recorded internally in
QDoc's `Node` representation of documentable entities.

Generation of this status is different for C++ class pages and QML type
pages, in that, if an entire class is deprecated, the C++ class page
generates a deprecation notice in the header, directly following the
"requisites" section (include statements, build system commands, etc).
QML type pages, while featuring the same "requisites" section, does not
generate the QML type status there.

This patch unifies this behavior by ensuring the responsible methods
for generating QML type pages also generate the status for the `Node`
QDoc is generating documentation for. This patch thus modifies both the
`HTMLGenerator` and `DocBookGenerator` classes. Updated test data for
the generated output test is included. As QML module versions may differ
from Qt version numbers, the warning text is modified for QML entities
such that it includes the QML module name for entites that belong to a
QML module.

Task-number: QTBUG-114157
Change-Id: I139b9f7f278f0d21fe82f3520cc095fa2e5e34fe
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>