qdoc: Use QML module information when searching for QML types
Use the findQmlTypeInPrimaryTree() function from QDocDatabase to
locate existing QML type nodes when processing \qmltype or
\qmlproperty/method/signal commands.

Not only do we want to restrict searching to the primary tree
(current documentation set), but we also want to use the
QML module information in the search. At the time of
processing a \qmltype topic, the module information is not yet
recorded in the Node instance. Therefore, look ahead for
the COMMAND_INQMLMODULE (\inqmlmodule) argument in the
set of meta-commands for the current topic.

Likewise for \qmlproperty (et al) topics, switch to searching
in the primary tree only. If the \qmlproperty argument includes
a module identifier, the parent QML node *must* be from that
module.

Note: This marks a change in QDoc's behavior; previously,
      using an incorrect (or non-existent) module identifier,
      as in

      \qmlproperty bool Whatever::Item::enabled

     still produced correct results, if a node with the correct
     type name existed in any module.

If the \qmlproperty command is processed before the associated
\qmltype documentation, QDoc constructs a placeholder
QmlTypeNode. Add this constructed node to the correct module so
subsequent searches will find it.

Apply the same search logic to QmlDocVisitor as well. Delay the
placeholder QmlTypeNode construction to happen in
QmlDocVisitor::applyDocumentation() implementation, so that
we can dig out the potential \inqmlmodule argument from the
Doc instance. Change the return value of the function to
be the (constructed) node pointer; the previous boolean
return value wasn't used for anything meaninful.

These changes improve QDoc's behavior on incorrect/ambiguous
module information passed to QML-related topic commands, and
allow a specific use case of documenting multiple QML modules
that contain identical type names in the same documentation
project. Related improvements to QDoc warning generation and
to the contents of QDoc manual are still needed.

On passing, clean up QmlDocVisitor::applyDocumentation():

 * Return early on invalid comment location; eliminate
   the long conditional block
 * Eliminate construction of multiple Location instances,
   use only one
 * Eliminate unnecessary call to QString::mid()

Fixes: QTBUG-114057
Change-Id: I57174c622ac64e0e0865b52887f15f4050e51a93
Reviewed-by: Paul Wicking <paul.wicking@qt.io>