qdoc: Output warnings for QML docs with incorrect module identifier
QDoc generates placeholder QmlTypeNode and CollectionNode instances when
it comes across an \inqmlmodule command whose argument refers to a
yet-to-be-seen \qmlmodule page.

The same happens when a \qmlproperty|method\signal argument refers
an unknown module identifier.

The purpose of these placeholders is to allow random ordering of
\qmltype, \qmlmodule, and \qmlproperty documentation.

When QDoc reaches the state of generating output, all these
placeholders should be now 'filled in' with their documentation.
Problems arise when the author forgets to document a \qmlmodule or
is using an incorrect module identifier for \inqmlmodule or
\qmlproperty commands; now, there are placeholder nodes in the
tree without documentation. A QmlTypeNode placeholder can
contain documented properties, but because the type itself
remains undocumented, its reference page is not generated.

In CppCodeParser, set location for placeholder QML type nodes.
This is the location of the property/method being processed,
and will be replaced with the location of the actual type once
it is 'seen'. This temporary location is used for warnings
when a type remains in the tree as a placeholder.

In Generator::generateDocumentation(), output warnings for QML types in
remaining QML modules that are undocumented ('unseen' by QDoc).

In QmlDocVisitor::applyDocumentation(), remove redundant check for
a valid topic command. The list of possible topic commands is
governed by QmlCodeParser; any topic not recognized by the QML
parser already results in 'Unknown command' warning.

Change-Id: I8fdeb4e1efb8be83398cc5d90fb28ddb75d8a0a4
Reviewed-by: Paul Wicking <paul.wicking@qt.io>